requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
jinja2==3.1.4

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.30.0
aiosqlite==0.20.0

# Data serialization
pydantic==2.8.2
python-multipart==0.0.7

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Env vars & utilities
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0

# Dev tools
pytest==8.3.2
yamllint==1.35.1